trade_backend/requirements.txt
# Django Core
Django>=5.2.8
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database (PyMySQL is used via config/__init__.py)
# mysqlclient removed - using PyMySQL instead

# Environment
python-dotenv>=1.0.0

# AI/ML
openai>=2.8.1
openai-agents>=0.6.1
langchain-core>=1.1.0
langchain-openai>=1.1.0
langchain-text-splitters>=1.0.0
tavily-python>=0.7.13
langfuse>=2.0.0

# Data Processing (already installed)
pandas>=2.3.3
numpy>=2.3.5
openpyxl>=3.1.5

# Vector Database (already installed)
qdrant-client>=1.16.1
mem0ai>=0.1.0

# AWS
boto3>=1.35.0

# PDF Processing
pymupdf>=1.24.0

# Utils (already installed)
PyMySQL>=1.1.2
requests>=2.32.5
python-multipart>=0.0.20
